Autorise les quantités décimales partout (achat, vente, devis, stock)
Troisième et dernière étape de l'ouverture hors zone FCFA (après devise et
TVA) : toutes les colonnes de quantité transactionnelle passent de entier à
decimal(12,3) — stock, mouvements de stock, lignes de vente/achat/devis/
vente-en-attente, retours, inventaire, transferts, bons de livraison.
unite_ventes.facteur reste volontairement entier (multiplicateur d'un lot,
pas une quantité). Saisie toujours autorisée en décimal, sans bascule dans
Paramètres — une quantité entière garde un comportement et un affichage
strictement identiques à avant (helper quantite(), trim des zéros non
significatifs). Un point ou une virgule sont acceptés en saisie, toujours
normalisés au point avant validation/stockage (App\Support\Decimal).

Points de casse corrigés au passage : casts Eloquent 'integer' → 'decimal:3'
sur toutes les colonnes concernées, signatures de service élargies en
int|float (StockService, TransfertService, InventaireService), comparaisons
strictes === devenues risquées sur des valeurs désormais castées en chaîne
(LigneVente::prixUnitaireEffectif(), écart d'inventaire), calculs
prix × quantité repassés par Arrondi::entier() pour rester des montants
entiers exacts. Complète aussi deux oublis du remplacement montant() de la
phase Devise, découverts en touchant ces mêmes fichiers (exports PDF/Excel
de RapportController et consorts qui gardaient encore " F" en dur).

Testé de bout en bout : migration + rollback + remigration sans perte de
données, suite de 10 scénarios métier via tinker (réception, vente avec/
sans unité de vente à facteur, achat, transfert, inventaire à écart négatif,
retour client — transaction annulée), et parcours HTTP réel (saisie virgule
"1,5" acceptée, ticket affichant "1.5" proprement, régression nulle sur une
quantité entière).

+++ app/Http/Controllers/BonLivraisonController.php
@@ -28,7 +28,7 @@ class BonLivraisonController extends Controller
             'motif' => ['nullable', 'string', 'max:500'],
             'lignes' => ['required', 'array', 'min:1'],
             'lignes.*.ligne_vente_id' => ['required', Rule::exists('ligne_ventes', 'id')->where('vente_id', $vente->id)],
-            'lignes.*.quantite_pieces' => ['required', 'integer', 'min:0'],
+            'lignes.*.quantite_pieces' => ['required', 'numeric', 'min:0'],
         ]);
 
         try {
